Post merge dev

+++ actions/deleteCoverLetter.ts
@@ -1,24 +1,25 @@
 "use server";
-import { PrismaClient } from "@prisma/client";
 import { auth } from "@/auth";
+import { asyncHandler } from "@/lib/actionsHelpers/actionsAsyncHandler";
+import { ActionsError } from "@/lib/actionsHelpers/actionsErrorHandler";
+import { prisma } from "@/prisma";
+import { revalidatePath } from "next/cache";
 
-const prisma = new PrismaClient();
-
-export async function deleteCoverLetter(coverId: string) {
-  console.log("reached delete");
+export const deleteCoverLetter = asyncHandler(async (coverId: string) => {
   const session = await auth();
-  try {
-    await prisma.coverletter.deleteMany({
-      where: {
-        id: coverId,
-        userId: session?.user?.id,
-      },
-    });
-    return { success: true, message: "Successfully deleted the cover letter" };
-  } catch (error) {
-    console.error("Error deleting cover letter:", error);
-    throw new Error("Error deleting the cover letter");
-  } finally {
-    await prisma.$disconnect();
-  }
-}
+  if (!session || !session?.user?.id) throw ActionsError.userNotAuthenticated;
+  if (!coverId) throw ActionsError.badRequest;
+
+  await prisma.coverletter.deleteMany({
+    where: {
+      id: coverId,
+      userId: session?.user?.id,
+    },
+  });
+  revalidatePath("/home", "page");
+  return {
+    success: true,
+    message: "Successfully deleted the cover letter",
+    status: 200,
+  };
+});
