+++ actions/deleteResume.ts
@@ -1,24 +1,24 @@
 "use server";
-import { PrismaClient } from "@prisma/client";
 import { auth } from "@/auth";
+import { asyncHandler } from "@/lib/actionsHelpers/actionsAsyncHandler";
+import { ActionsError } from "@/lib/actionsHelpers/actionsErrorHandler";
+import { prisma } from "@/prisma";
+import { revalidatePath } from "next/cache";
 
-const prisma = new PrismaClient();
-
-export async function deleteResume(resumeId: string) {
-    console.log("reached delete")
+export const deleteResume = asyncHandler(async (resumeId: string) => {
   const session = await auth();
-  try {
-    await prisma.resume.deleteMany({
-      where: {
-        id: resumeId,
-        userId: session?.user?.id,
-      },
-    });
-    return { success: true, message: "Successfully deleted the resume" }
-  } catch (error) {
-      console.error("Error deleting resume:", error);
-      throw new Error("Error deleting the resume");
-  } finally {
-      await prisma.$disconnect();
-  }
-}
+  if (!session || !session?.user?.id) throw ActionsError.userNotAuthenticated;
+  if (!resumeId) throw ActionsError.badRequest;
+  await prisma.resume.deleteMany({
+    where: {
+      id: resumeId,
+      userId: session?.user?.id,
+    },
+  });
+  revalidatePath("/home", "page");
+  return {
+    success: true,
+    message: "Successfully deleted the resume",
+    status: 200,
+  };
+});
